// MS_rosterfrontend/projects/client/src/lib/view-models/client-basicinfo.ts
export interface ClientBasicData {
    clientId?: number;
    salutation?: string;
    firstName?: string;
    middleName?: string;
    lastName?: string;
    dateOfBirth: string;
    mobileNo?: string;
    locationId: number;
    emailId?: string;
    address?: number;
    fullName?: string;
    age?: number;
    gender?: string;
    genderName?:string;
    locationName?:string; 
    salutationName?:string;
    serviceType?: number;
    ndis?: string;
    imageUrl?: string;
    otherLocation?:string; 
    locationType?:number;
    status?:boolean;
    userName? :string;
}